extensively use mountComponent utils
