+++ lib/components/result-view/index.js
@@ -4,6 +4,7 @@ import React from "react"
 import ReactDOM from "react-dom"
 
 import ResultViewComponent from "./result-view"
+import {mountComponent} from "../../utils"
 
 export default class ResultView {
   destroy = () => {
@@ -12,17 +13,14 @@ export default class ResultView {
 
   constructor(outputStore, kernel, showResult = true) {
     this.outputStore = outputStore
-    this.element = document.createElement("div")
-    this.element.classList.add("hydrogen", "marker")
-
-    ReactDOM.render(
+    this.element = mountComponent(
       <ResultViewComponent
         store={outputStore}
         kernel={kernel}
         destroy={this.destroy}
         showResult={showResult}
       />,
-      this.element
+      ["marker"]
     )
   }
 }
