+++ lib/hydrogen-pane.js
@@ -12,20 +12,7 @@ import ReactDOM from "react-dom"
 import Watches from "./components/watch-sidebar"
 import Inspector from "./components/inspector"
 import OutputArea from "./components/output-area"
-
-// Create div container element and render to container
-// Return container element
-function mountComponent(component, containerClassList = []) {
-  const element = document.createElement("div")
-
-  element.classList.add("hydrogen")
-  if (containerClassList.length) {
-    element.classList.add(...containerClassList)
-  }
-  ReactDOM.render(component, element)
-
-  return element
-}
+import {mountComponent} from "./utils"
 
 class HydrogenPane {
   static isOpenable(uri) {
