+++ lib/main.js
@@ -19,7 +19,7 @@ import Config from "./config"
 import kernelManager from "./kernel-manager"
 import ZMQKernel from "./zmq-kernel"
 import AutocompleteProvider from "./autocomplete-provider"
-import {renderMobxReactDevtools, hotReloadPackage} from "./utils"
+import {renderMobxReactDevtools, hotReloadPackage, mountComponent} from "./utils"
 
 import HydrogenPane from "./hydrogen-pane"
 
@@ -119,11 +119,8 @@ export default {
   },
 
   consumeStatusBar(statusBar) {
-    const element = document.createElement("div")
-    element.className = "inline-block"
     const onClick = this.showKernelCommands.bind(this)
-    ReactDOM.render(<StatusBar store={store} onClick={onClick} />, element)
-
+    const element = mountComponent(<StatusBar store={store} onClick={onClick} />, ["inline-block"])
     const tile = statusBar.addLeftTile({item: element, priority: 100})
     this.subscriptions.add(new Disposable(() => tile.destroy()))
   },
