+++ lib/store/watches.js
@@ -15,7 +15,7 @@ export default class WatchesStore {
     this.addWatch()
   }
 
-  @action createWatch = () => {
+  @action createWatch() {
     const lastWatch = this.watches[this.watches.length - 1]
     if (!lastWatch || lastWatch.getCode().replace(/\s/g, "") !== "") {
       const watch = new WatchStore(this.kernel)
