+++ lib/utils.js
@@ -95,10 +95,9 @@ export function log(...message) {
 
 export function renderMobxReactDevtools() {
   const devTools = require("mobx-react-devtools")
-  const div = document.createElement("div")
-  document.getElementsByTagName("body")[0].appendChild(div)
+  const element = mountComponent(<devTools.default noPanel />)
+  document.getElementsByTagName("body")[0].appendChild(element)
   devTools.setLogEnabled(true)
-  ReactDOM.render(<devTools.default noPanel />, div)
 }
 
 export function deprecationNote() {
@@ -183,6 +182,20 @@ export function limitNumber(number, {min, max} = {}) {
   if (min != null) number = Math.max(number, min)
   if (max != null) number = Math.min(number, max)
   return number
+}
+
+// Create div container element and render to container
+// Return container element
+export function mountComponent(component, containerClassList = []) {
+  const element = document.createElement("div")
+
+  element.classList.add("hydrogen")
+  if (containerClassList.length) {
+    element.classList.add(...containerClassList)
+  }
+  ReactDOM.render(component, element)
+
+  return element
 }
 
 // WIP
